requirements.txt
django
bleach
markdown
django-simple-captcha
argon2-cffi
pyotp
qrcode
pillow
gunicorn
django-environ
cryptography